Revert "Merge branch 'master' of bitbucket.org:spectrometerH/ttrs"
This reverts commit dea9443c06e07d8fa826c76b839aa889a7312772, reversing
changes made to e0f63bfae26c9cd282104f9479a40f629180b0ce.

+++ backend/UserManager.hpp
@@ -15,7 +15,7 @@ private:
 		Password password;
 		Email email;
 		Phone phone;
-		short privilege;
+		int privilege;
 		User() : privilege(1) {}
 	};
